requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
google-generativeai==0.3.1
plotly==5.18.0
numpy==1.26.2
python-multipart==0.0.6
cassandra-driver==3.28.0
python-jose[cryptography]
passlib[bcrypt]
